Scoped Timer + fp precision
ScopedTimer api added
getTime() and sleepFor now use floats instead of doubles

+++ Timer.cpp
@@ -4,9 +4,9 @@ using namespace std::chrono;
 
 high_resolution_clock::time_point appStartTime = high_resolution_clock::now();
 
-double getTime() {
+float getTime() {
 
 	high_resolution_clock::time_point now = high_resolution_clock::now();
-	return duration<double, std::chrono::seconds::period>(now - appStartTime).count(); 
+	return duration<float, std::chrono::seconds::period>(now - appStartTime).count(); 
 
 }
